refactor: split logging functions away from config
